backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
httpx
python-dotenv